change: TextureCoordinate -> TexCoord_Color; feat_begin: Numerous new features in Scene, World and Entity

+++ VolcaniCore/src/VolcaniCore/Core/Random.cpp
@@ -4,14 +4,14 @@ namespace VolcaniCore {
 
 static std::random_device s_Device;
 static std::mt19937 s_RNG{ s_Device() };
-static std::uniform_int_distribution<uint32_t> s_Dist{ };
+static std::uniform_int_distribution<uint64_t> s_Dist{ };
 
 uint32_t Random::RandInt() {
 	return s_Dist(s_RNG);
 }
 
 uint32_t Random::RandInt(uint32_t low, uint32_t high) {
-	std::uniform_int_distribution<uint32_t> dist{ low, high };
+	std::uniform_int_distribution<uint64_t> dist{ low, high };
 	return dist(s_RNG);
 }
